Fix runUnitTest for Zope 2.13 by disabling ZopeLite patches
A ZopeLite patch was added to prevent loading of zcml files, which is wrong
for ERP5: CMF products were not properly initialized.

Code is also simpler now that we call OFS.Application.initialize(), which is
not significantly slower.

We keep patch not to register help though because this is quite expensive:
we should even consider always disable help because they seems to cause many
conflicts during Zope cluster startup.

A monkey-patch to unittest is updated for Python 2.7, so that tests aren't run
when they shouldn't (e.g. --save without --load).

Conflicts:
	product/ERP5OOo/tests/testOOoImport.py