PythonScript, SQLMethod: Fix expression_instance getting stale
Before this change, expression_instance would not be set when expression is
modified, leading to either non-applied filter expression (when no
expression_instance existed) or use of an outdated expression.
Use a volatile attribute to store the expression: it does not change the
amount of work needed to load the Expression instance (pickle is the raw
python expression as text) and it remove data duplication in exported
documents.